refactor(agent): fold OutputFileProcessor into outputFileExtraction
`OutputFileProcessor` was a single-use class: one production construction
site (`outputFileExtraction.ts`), immediately discarded after one method
call, with three constructor fields that were exactly the three parameters
the surrounding module function already had in hand.

Fold it into `outputFileExtraction.ts` as module functions and delete the
file. `emitMissingOutputs` had a single caller (`handleNoOutputs`) and is
merged into it.

Behavior is unchanged.
